requirements.txt
addict==2.4.0
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1725434011349/work
aiohttp @ file:///croot/aiohttp_1725527756643/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
async-timeout @ file:///croot/async-timeout_1703096998144/work
attrs @ file:///croot/attrs_1729089401488/work
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///home/task_175978575413734/conda-bld/certifi_1759786089495/work/certifi
cftime==1.6.4.post1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
contourpy @ file:///croot/contourpy_1738160616259/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
decorator==5.2.1
docker-pycreds==0.4.0
et_xmlfile @ file:///croot/et_xmlfile_1756462486635/work
filelock @ file:///croot/filelock_1700591183607/work
fonttools==4.54.1
frozenlist @ file:///croot/frozenlist_1698702560391/work
fsspec @ file:///croot/fsspec_1724855583730/work
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
idna @ file:///croot/idna_1714398848350/work
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib @ file:///croot/joblib_1718217211762/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
mmcv==1.3.11
mmengine==0.10.5
mne==1.8.0
mne-connectivity==0.7.0
moviepy==2.2.1
mpmath @ file:///croot/mpmath_1690848262763/work
multidict @ file:///croot/multidict_1701096859099/work
netCDF4==1.7.2
networkx @ file:///croot/networkx_1717597493534/work
numexpr @ file:///croot/numexpr_1730215937391/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0
opencv-python-headless==4.10.0
openpyxl @ file:///croot/openpyxl_1736366913861/work
packaging==24.1
pandas @ file:///home/task_175982151591391/conda-bld/pandas_1759822246465/work/dist/pandas-2.3.3-cp39-cp39-linux_x86_64.whl#sha256=46f49f8c0e38edb70c0c31fa0bcb595d56fee69beeb19b55154de62090265ed3
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.12
protobuf==5.28.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
pyg-lib==0.4.0+pt26cu124
Pygments==2.18.0
pyparsing @ file:///croot/pyparsing_1725041624202/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.2.1
pytz @ file:///croot/pytz_1752135852232/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests @ file:///croot/requests_1721410876868/work
rich==13.9.4
scikit-image==0.24.0
scikit-learn @ file:///croot/scikit-learn_1721921875708/work
scipy @ file:///croot/scipy_1717521478074/work/dist/scipy-1.13.1-cp39-cp39-linux_x86_64.whl#sha256=2cb3d1436f8762a53161486cb9493ccc305fc760e8f830f82888792cbaf99099
seaborn @ file:///croot/seaborn_1749110291192/work
sentry-sdk==2.18.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.1
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tifffile==2024.8.30
tomli==2.0.2
torch==2.6.0+cu124
torch-geometric==2.6.1
torch_cluster==1.6.3+pt26cu124
torch_scatter==2.1.2+pt26cu124
torch_sparse==0.6.18+pt26cu124
torch_spline_conv==1.2.2+pt26cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm @ file:///croot/tqdm_1724853939799/work
triton==3.2.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3 @ file:///croot/urllib3_1727769808118/work
xarray==2024.7.0
yapf==0.40.2
yarl @ file:///croot/yarl_1725976495189/work
zipp==3.20.2
